+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 Improvements
 
 - Added a component to render Hooklifts logo.
+- hlInput autocomplete is disabled by default.